pyproject.toml
[project]
name = "demo-foundry-iq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-agents>=1.2.0b6",
    "azure-ai-projects>=2.0.0b4",
    "azure-identity>=1.25.0",
    "azure-mgmt-cognitiveservices>=14.0.0",
    "azure-search-documents>=11.7.0b2",
    "azure-storage-blob>=12.20.0",
    "openai>=1.40.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "rich>=13.7.0",
]
